Initial commit
It's currently just a mess of models and services all being used from
the main activity. When run, the app just displays a button that can be
clicked to recieve a random speedrun and display a few bits of
information about it. It logs a weblink to the video but cannot be
viewed inside the app or emulator just yet.