Added ViewRunActivity
Added a new activity to view specific run information and watch a video of it. Used the embedded YouTube player. Hardcoded in a lot of sample data for the ViewRunActivity and ViewLeaderboardActivity just for this part of my assignment. Will be overhauling this with RxJava requests for real data soon.
Built a decent looking UI that I'm pleased with for now. I'll have to revisit some things like the title in ViewRunActivity as it doesn't make sense when accessed with the random button, and the favourite menu option in ViewRunActivity has a different meaning than intended.