Implementing Shared Prefs
Implemented shared preferences for storing the games list and favourites. I made the read/write methods generic so they can handle an type of list but at the moment it actually only handles lists of GameModel.

The first time the app is run, the game list is parsed from the raw json file included in the apk and stored to be reused. Honestly this approach is awful and I should revisit it later to just make small requests for the first hundred or so games