Hooked up the networking
The categories and leaderboards are properly populating with requests to the API now. The run activity now displays different runs and either displays the video in the Youtube player if it is a Youtube link, or allows the user to start up another app to watch links from other websites.

Big problem is that storing the entire gameList uses up WAY more space than I thought it would. I might just continue to parse the raw json every time and use storage for favourites only.