Displaying multiple runners
To allow the app to display multiple runners I had to cut the links to the runner channels due to time constraints. When I have time to revisit the UI I will reimplement this so I can display multiple runner links.

Changed favourites so they are saved as just Strings rather than full GameModel objects. This is both to save space when saving to disk and also to help fix a bug where the favourites list would display game titles but no platform names