Fixes
There was a crash caused by a 404 for a random request. I added an errorConsumer to log the error. It seems extremely unlikely to happen but if it does, it will just silently fail for now.

Updated the time formatter to give a better looking time string. It was very ambiguous when it was just seconds